ml-model/requirements.txt
# ============================================================================
# MINIMAL ESSENTIALS - Only What's Actually Needed
# ============================================================================

# Core ML (Just the 3 ensemble models we use)
xgboost==2.0.3
lightgbm==4.3.0
scikit-learn==1.4.0

# Data Processing (Required)
numpy==1.26.3
pandas==2.1.4

# URL Analysis (Required)
tldextract==5.1.1
dnspython==2.4.2
python-whois==0.8.0
python-Levenshtein==0.25.0

# SSL Analysis (Required)
cryptography==41.0.7
pyOpenSSL==23.3.0
certifi==2023.11.17

# Web Scraping (Required)
requests==2.31.0
beautifulsoup4==4.12.2
lxml==5.1.0

# Utilities (Required)
tqdm==4.66.1
joblib==1.3.2
colorama==0.4.6
onnx==1.15.0
onnxruntime==1.16.3
tf2onnx==1.16.1

# Explainability & Interpretability
shap==0.44.1
lime==0.2.0.1

# Experiment Tracking & Monitoring
mlflow==2.9.2
tensorboard==2.15.1
wandb==0.16.2

# Performance & Profiling
memory-profiler==0.61.0
py-spy==0.3.14
line-profiler==4.1.1

